{{cookiecutter.project_slug}}/requirements-test.txt
{%- if cookiecutter.use_pytest == 'y' %}pytest==3.6.1
pytest-runner==4.2      # add setup.py test support for pytest
{%- endif %}
